Extract BaseItemRegistry into its own base_item_registry.py file
